refactor: rename `appDirectory` to `src/`
