+++ apps/react-router-ssr/react-router.config.ts
@@ -2,5 +2,6 @@ import { type Config } from '@react-router/dev/config';
 
 export default {
   ssr: true,
+  appDirectory: 'src',
   prerender: ['/about'],
 } satisfies Config;
