+++ apps/react-router-ssr/src/components/LoadingSplash/index.tsx
