+++ apps/react-router-ssr/src/entry.client.tsx
@@ -1,0 +1,10 @@
+import { StrictMode } from 'react';
+import { hydrateRoot } from 'react-dom/client';
+import { HydratedRouter } from 'react-router/dom';
+
+hydrateRoot(
+  document,
+  <StrictMode>
+    <HydratedRouter />
+  </StrictMode>,
+);
