+++ apps/react-router-ssr/src/routes.ts
