+++ apps/react-router-ssr/src/routes/_app._index/route.tsx
