+++ apps/react-router-ssr/src/routes/_app.contacts.$contactId/Template/Favorite/index.tsx
