+++ apps/react-router-ssr/src/routes/_app.contacts.$contactId/Template/Favorite/styles.module.css.d.ts
