+++ apps/react-router-ssr/src/routes/_app.new/route.tsx
