+++ apps/react-router-ssr/src/templates/ErrorPage/index.tsx
