+++ apps/react-router-ssr/src/templates/ErrorPage/styles.module.css.d.ts
